+++ packages/docs/src/components/logo/index.ts
@@ -1,9 +1,0 @@
-import { Logo } from './class';
-import { styles } from './styles';
-import { logoTemplate as template } from './template';
-
-export const SdsLogo = Logo.compose({
-  baseName: 'logo',
-  styles,
-  template,
-});