+++ packages/docs/src/components/logo/options.ts
@@ -1,4 +1,0 @@
-/**
- * The logo is displayed in original colours. It should ideally be placed on a white background
- */
-export type LogoVariant = 'black' | 'white' | 'color';