+++ packages/docs/CHANGELOG.md
@@ -34,3 +34,4 @@ The following icons are used in commit messages and this changelog.
 ### ⛔ Removals
 
 - #20: Removed Storybook demo content
+- #18: Moved `<SDSLogo />` component to `@sick-design-system/components`